+++ lib/activeadmin_addons/addons/toggle_bool_builder.rb
@@ -24,7 +24,9 @@ module ActiveAdminAddons
         'data-field' => attribute,
         'data-value' => data,
         'data-url' => resource_url,
-        'data-success_message' => options[:success_message]
+        'data-success_message' => options[:success_message],
+        'data-confirm_message' => options[:confirm_message],
+        'data-confirm_message_trigger' => options[:confirm_message_trigger]
       )
     end